Update MontosFinancieros
Actualizados los tipos de datos de montos financieros a decimal tanto en la base de datos como en la aplicación para el mejor manejo de los mismos. Correcciones necesarias en el programa para facilitar los cambios realizados. Errores detectados corregidos.

+++ aDVanceERP.Modulos.CompraVenta/MVP/Modelos/Compra.cs
@@ -5,7 +5,7 @@ namespace aDVanceERP.Modulos.CompraVenta.MVP.Modelos {
         public Compra() {
         }
 
-        public Compra(long id, DateTime fecha, long idAlmacen, long idProveedor, float total) {
+        public Compra(long id, DateTime fecha, long idAlmacen, long idProveedor, decimal total) {
             Id = id;
             Fecha = fecha;
             IdAlmacen = idAlmacen;
@@ -17,7 +17,7 @@ namespace aDVanceERP.Modulos.CompraVenta.MVP.Modelos {
         public DateTime Fecha { get; set; }
         public long IdAlmacen { get; set; }
         public long IdProveedor { get; set; }
-        public float Total { get; set; }
+        public decimal Total { get; set; }
     }
 
     public enum CriterioBusquedaCompra {
