+++ aDVanceERP.Modulos.CompraVenta/MVP/Vistas/Pago/Plantillas/IVistaGestionPagos.cs
@@ -3,6 +3,6 @@ using aDVanceERP.Core.MVP.Vistas.Plantillas;
 
 namespace aDVanceERP.Modulos.CompraVenta.MVP.Vistas.Pago.Plantillas {
     public interface IVistaGestionPagos : IVistaContenedor, IGestorDatos {
-        void AdicionarPago(string metodoPago = "", float monto = -1f);
+        void AdicionarPago(string metodoPago = "", decimal monto = -1);
     }
 }
