+++ aDVanceERP.Modulos.CompraVenta/MVP/Vistas/Pago/Plantillas/IVistaRegistroPago.cs
@@ -3,12 +3,12 @@
 namespace aDVanceERP.Modulos.CompraVenta.MVP.Vistas.Pago.Plantillas {
     public interface IVistaRegistroPago : IVistaRegistro {
         string MetodoPago { get; set; }
-        float Monto { get; set; }
+        decimal Monto { get; set; }
         List<string[]> Pagos { get; }
-        float Total { get; set; }
-        float Suma { get; set; }
-        float Pendiente { get; set; }
-        float Devolucion { get; set; }
+        decimal Total { get; set; }
+        decimal Suma { get; set; }
+        decimal Pendiente { get; set; }
+        decimal Devolucion { get; set; }
 
         event EventHandler? EfectuarTransferencia;
         event EventHandler? PagoAgregado;
